layers: Move struct interface_var to simplify includes
